Only use useCaseUtils to access useCaseRelevant data within the components
