+++ webofneeds/won-owner-webapp/src/main/webapp/app/usecase-utils.js
@@ -152,6 +152,160 @@ export function getAllMessageDetails() {
   return allMessageDetails;
 }
 
+/**
+ * returns an object containing all use cases that were
+ * not found in a useCaseGroup or in a group that's too
+ * small to be displayed as a group
+ * @param threshold => defines size that is deemed too small to display group
+ * @returns {*}
+ */
+export function getUnGroupedUseCases(threshold = 0) {
+  const useCaseGroups = getUseCaseGroups();
+  let ungroupedUseCases = JSON.parse(JSON.stringify(getUseCases()));
+
+  for (const identifier in useCaseGroups) {
+    const group = useCaseGroups[identifier];
+    // show use cases from groups that can't be displayed
+    // show use cases from groups that have no more than threshold use cases
+    if (
+      !isDisplayableUseCaseGroup(group) ||
+      countDisplayableUseCasesInGroup(group) <= threshold
+    ) {
+      continue;
+    }
+    // don't show usecases in groups as sinle use cases
+    for (const useCase in group.useCases) {
+      if (group.useCases[useCase].useCases) {
+        for (const subUseCase in group.useCases[useCase].useCases) {
+          //FIXME: SUBSUB GROUPS ARE NOT SUPPORTED YET
+          delete ungroupedUseCases[subUseCase];
+        }
+      } else {
+        delete ungroupedUseCases[useCase];
+      }
+    }
+  }
+  return ungroupedUseCases;
+}
+
+export function getUseCase(useCaseString) {
+  if (useCaseString) {
+    const useCases = getUseCases();
+
+    for (const useCaseName in useCases) {
+      if (useCaseString === useCases[useCaseName]["identifier"]) {
+        return useCases[useCaseName];
+      }
+    }
+  }
+  return undefined;
+}
+
+/**
+ * This fucntion returns all UseCases that are defined with "showInList = true", this is so we can define useCases
+ * that show up within our connection-overview "quick item"-buttons, e.g. if we want to define a usecase that is
+ * always available from this list
+ * @returns {{}}
+ */
+export function getListUseCases() {
+  let listUseCases = {};
+  const useCases = getUseCases();
+
+  for (const useCaseKey in useCases) {
+    if (useCases[useCaseKey]["showInList"]) {
+      listUseCases[useCaseKey] = useCases[useCaseKey];
+    }
+  }
+  return listUseCases;
+}
+
+export function getUseCaseGroupByIdentifier(groupIdentifier) {
+  if (groupIdentifier) {
+    const useCaseGroups = getUseCaseGroups();
+    for (const groupName in useCaseGroups) {
+      const element = useCaseGroups[groupName];
+
+      if (groupIdentifier === element["identifier"]) {
+        return element;
+      } else if (isUseCaseGroup(element)) {
+        for (const subGroupName in element.useCases) {
+          // FIXME: DOES NOT WORK FOR SUBSUB GROUPS
+          if (
+            groupIdentifier === element.useCases[subGroupName]["identifier"]
+          ) {
+            return element.useCases[subGroupName];
+          }
+        }
+      }
+    }
+  }
+  return undefined;
+}
+
+/**
+ * return if the given useCase is displayable or not
+ * @param useCase
+ * @returns {*}
+ */
+export function isDisplayableUseCase(useCase) {
+  return useCase && useCase.identifier && (useCase.label || useCase.icon);
+}
+
+/**
+ * return if the given useCaseGroup is displayable or not
+ * needs to have at least one displayable UseCase
+ * @param useCase
+ * @returns {*}
+ */
+export function isDisplayableUseCaseGroup(useCaseGroup) {
+  const useCaseGroupValid =
+    useCaseGroup &&
+    (useCaseGroup.label || useCaseGroup.icon) &&
+    useCaseGroup.useCases;
+
+  if (useCaseGroupValid) {
+    for (const key in useCaseGroup.useCases) {
+      if (isDisplayableUseCase(useCaseGroup.useCases[key])) {
+        return true;
+      }
+    }
+  }
+  return false;
+}
+
+/**
+ * return the amount of displayable useCases in a useCaseGroup
+ * @param useCaseGroup
+ * @return {*}
+ */
+export function countDisplayableUseCasesInGroup(useCaseGroup) {
+  let countUseCases = 0;
+
+  for (const key in useCaseGroup.useCases) {
+    if (isDisplayableUseCase(useCaseGroup.useCases[key])) {
+      countUseCases++;
+    }
+  }
+  return countUseCases;
+}
+
+/**
+ * return if the given element is a useCaseGroup or not
+ * @param element
+ * @returns {*}
+ */
+export function isUseCaseGroup(element) {
+  return element.useCases;
+}
+
 function hasSubElements(obj) {
   return obj && obj !== {} && Object.keys(obj).length > 0;
+}
+
+export function getUseCaseGroups() {
+  return useCaseDefinitions.getUseCaseGroups();
+}
+
+export function getUseCases() {
+  return useCaseDefinitions.getUseCases();
 }
